+++ packages/syllst-compare/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @syllst/compare
 
+## 0.2.1
+
+### Patch Changes
+
+- update config
+- Updated dependencies
+  - @syllst/core@0.6.1
+
 ## 0.2.0
 
 ### Minor Changes