+++ packages/syllst-content-shared/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @syllst/content-shared
 
+## 1.0.2
+
+### Patch Changes
+
+- update config
+- Updated dependencies
+  - @syllst/processor@0.6.1
+  - @syllst/core@0.6.1
+
 ## 1.0.1
 
 ### Patch Changes