+++ packages/syllst-glost/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @syllst/glost
 
+## 0.5.6
+
+### Patch Changes
+
+- update config
+- Updated dependencies
+  - @syllst/core@0.6.1
+
 ## 0.5.2
 
 ### Patch Changes