+++ packages/syllst-xapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @syllst/xapi
 
+## 0.4.5
+
+### Patch Changes
+
+- update config
+- Updated dependencies
+  - @syllst/core@0.6.1
+
 ## 0.4.4
 
 ### Patch Changes